IGNITE-17298 RecordMarshallerTest moved to tests.
